+++ users/src/lib/config.ts
@@ -42,9 +42,12 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
     let serverURL = options.serverURL;
 
     const params: Record<string, string> = {};
-    const serverIdx = options.serverIdx ?? 0;
 
     if (!serverURL) {
+        const serverIdx = options.serverIdx ?? 0;
+        if (serverIdx < 0 || serverIdx >= ServerList.length) {
+            throw new Error(`Invalid server index ${serverIdx}`);
+        }
         serverURL = ServerList[serverIdx] || "";
     }
 
@@ -55,7 +58,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = Object.freeze({
     language: "typescript",
     openapiDocVersion: "1.0.0",
-    sdkVersion: "2.0.2",
-    genVersion: "2.277.0",
-    userAgent: "speakeasy-sdk/typescript 2.0.2 2.277.0 1.0.0 @wingspan/users",
+    sdkVersion: "2.1.0",
+    genVersion: "2.281.2",
+    userAgent: "speakeasy-sdk/typescript 2.1.0 2.281.2 1.0.0 @wingspan/users",
 });