+++ users/RELEASES.md
@@ -39,3 +39,13 @@ Based on:
 - [typescript v2.0.2] users
 ### Releases
 - [NPM v2.0.2] https://www.npmjs.com/package/@wingspan/users/v/2.0.2 - users
+
+## 2024-03-18 01:27:07
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.wingspan.app/openapi/6470eea6a7577e003be20a26
+- Speakeasy CLI 1.209.3 (2.281.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v2.1.0] users
+### Releases
+- [NPM v2.1.0] https://www.npmjs.com/package/@wingspan/users/v/2.1.0 - users